Allow four satellite network
This contract is complete-able with either three or four satellites, but encourages using four satellites just like the tutorial. Because Commsat I and Commsat III can be the same probe, it only takes 2 satellites to complete the placement, but it will fail the 2 day test until the network is completed with a 3rd (or 4th) satellite.